将 VerticalRepeaterView 改为 TemplateControl 以提供更好的性能
