--- src/App/Controls/Common/VerticalRepeaterView.xaml
@@ -1,72 +1,0 @@
-﻿<UserControl
-    x:Class="Richasy.Bili.App.Controls.VerticalRepeaterView"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:Richasy.Bili.App.Controls"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
-    d:DesignHeight="300"
-    d:DesignWidth="400"
-    mc:Ignorable="d">
-    <UserControl.Resources>
-        <muxc:UniformGridLayout
-            x:Name="GridLayout"
-            ItemsStretch="Fill"
-            MinColumnSpacing="16"
-            MinItemHeight="{x:Bind MinWideItemHeight, Mode=OneWay}"
-            MinItemWidth="{x:Bind MinWideItemWidth, Mode=OneWay}"
-            MinRowSpacing="16"
-            Orientation="Horizontal" />
-
-        <muxc:StackLayout
-            x:Name="ListLayout"
-            Orientation="Vertical"
-            Spacing="8" />
-
-        <local:StaggeredLayout
-            x:Name="StaggeredLayout"
-            ColumnSpacing="16"
-            DesiredColumnWidth="{x:Bind MinWideItemWidth, Mode=OneWay}"
-            RowSpacing="16" />
-    </UserControl.Resources>
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-        <Grid Margin="0,0,0,8" Visibility="{x:Bind HeaderVisibility, Mode=OneWay}">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="Auto" />
-            </Grid.ColumnDefinitions>
-            <TextBlock
-                Style="{StaticResource SubtitleTextBlockStyle}"
-                HorizontalAlignment="Left"
-                VerticalAlignment="Center"
-                FontSize="16"
-                Text="{x:Bind HeaderText, Mode=OneWay}"
-                TextTrimming="CharacterEllipsis" />
-
-            <ContentPresenter
-                Grid.Column="1"
-                HorizontalAlignment="Right"
-                VerticalAlignment="Center"
-                HorizontalContentAlignment="Stretch"
-                VerticalContentAlignment="Center"
-                Content="{x:Bind AdditionalContent, Mode=OneWay}" />
-        </Grid>
-
-        <muxc:ItemsRepeater
-            x:Name="ItemsRepeater"
-            Grid.Row="1"
-            Margin="0,4,0,0"
-            ElementPrepared="OnElementPrepared"
-            ItemTemplate="{x:Bind ItemTemplate, Mode=OneWay}"
-            ItemsSource="{x:Bind ItemsSource, Mode=OneWay}"
-            Layout="{StaticResource GridLayout}"
-            VerticalCacheLength="2" />
-    </Grid>
-</UserControl>
