+++ src/App/Controls/Common/VerticalRepeaterView/VerticalRepeaterView.Properties.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Richasy. All rights reserved.
 
 using System;
-using System.Collections;
-using Microsoft.UI.Xaml.Controls;
-using Richasy.Bili.App.Resources.Extension;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 
@@ -12,7 +9,7 @@ namespace Richasy.Bili.App.Controls
     /// <summary>
     /// 视频视图.
     /// </summary>
-    public sealed partial class VerticalRepeaterView : UserControl
+    public sealed partial class VerticalRepeaterView
     {
         /// <summary>
         /// <see cref="ItemsSource"/>的依赖属性.
@@ -79,19 +76,6 @@ namespace Richasy.Bili.App.Controls
         /// </summary>
         public static readonly DependencyProperty IsStaggeredProperty =
             DependencyProperty.Register(nameof(IsStaggered), typeof(bool), typeof(VerticalRepeaterView), new PropertyMetadata(false, new PropertyChangedCallback(OnIsStaggeredChanged)));
-
-        private ScrollViewer _parentScrollViewer;
-        private double _itemHolderHeight = 0d;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="VerticalRepeaterView"/> class.
-        /// </summary>
-        public VerticalRepeaterView()
-        {
-            this.InitializeComponent();
-            this.Loaded += OnLoaded;
-            this.Unloaded += OnUnloaded;
-        }
 
         /// <summary>
         /// 在外部的ScrollViewer滚动到接近底部时发生.
@@ -195,135 +179,6 @@ namespace Richasy.Bili.App.Controls
         {
             get { return (bool)GetValue(IsStaggeredProperty); }
             set { SetValue(IsStaggeredProperty, value); }
-        }
-
-        private static void OnIsStaggeredChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            var instance = d as VerticalRepeaterView;
-            instance.CheckOrientationStatus();
-        }
-
-        private static void OnOrientationChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            var instance = d as VerticalRepeaterView;
-            if (e.NewValue is Orientation type)
-            {
-                instance.CheckOrientationStatus();
-            }
-        }
-
-        private void CheckOrientationStatus()
-        {
-            if (IsStaggered)
-            {
-                ItemsRepeater.Layout = StaggeredLayout;
-            }
-            else
-            {
-                switch (ItemOrientation)
-                {
-                    case Orientation.Vertical:
-                        ItemsRepeater.Layout = GridLayout;
-                        break;
-                    case Orientation.Horizontal:
-                        ItemsRepeater.Layout = ListLayout;
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            ChangeInitializedItemOrientation();
-        }
-
-        private void OnLoaded(object sender, RoutedEventArgs e)
-        {
-            if (EnableDetectParentScrollViewer)
-            {
-                _parentScrollViewer = this.FindAscendantElementByType<ScrollViewer>();
-                if (_parentScrollViewer != null)
-                {
-                    _parentScrollViewer.ViewChanged += OnParentScrollViewerViewChanged;
-                }
-            }
-        }
-
-        private void OnUnloaded(object sender, RoutedEventArgs e)
-        {
-            if (_parentScrollViewer != null)
-            {
-                _parentScrollViewer.ViewChanged -= OnParentScrollViewerViewChanged;
-                _parentScrollViewer = null;
-            }
-        }
-
-        private void OnParentScrollViewerViewChanged(object sender, ScrollViewerViewChangedEventArgs e)
-        {
-            if (!e.IsIntermediate && _parentScrollViewer != null)
-            {
-                var currentPosition = _parentScrollViewer.VerticalOffset;
-                if (_parentScrollViewer.ScrollableHeight - currentPosition <= _itemHolderHeight &&
-                    this.Visibility == Visibility.Visible)
-                {
-                    RequestLoadMore?.Invoke(this, EventArgs.Empty);
-                }
-            }
-        }
-
-        private void ChangeInitializedItemOrientation()
-        {
-            if (ItemsSource is ICollection items)
-            {
-                for (var i = 0; i < items.Count; i++)
-                {
-                    var element = ItemsRepeater.TryGetElement(i);
-                    if (element != null && element is IDynamicLayoutItem vi)
-                    {
-                        if (vi.Orientation != ItemOrientation)
-                        {
-                            vi.Orientation = ItemOrientation;
-                        }
-                    }
-                }
-            }
-        }
-
-        private void OnElementPrepared(ItemsRepeater sender, ItemsRepeaterElementPreparedEventArgs args)
-        {
-            if (args.Element != null)
-            {
-                if (args.Element is IDynamicLayoutItem dynamicLayoutItem)
-                {
-                    dynamicLayoutItem.Orientation = ItemOrientation;
-                }
-
-                if (IsAutoFillEnable &&
-                    args.Element is IRepeaterItem repeaterItem &&
-                    ItemsSource is ICollection collectionSource &&
-                    (_parentScrollViewer != null) &&
-                    args.Index >= collectionSource.Count - 1)
-                {
-                    var size = repeaterItem.GetHolderSize();
-                    _itemHolderHeight = size.Height;
-                    var viewportWidth = _parentScrollViewer.ViewportWidth;
-                    var viewportHeight = _parentScrollViewer.ViewportHeight;
-                    bool isNeedLoadMore;
-                    if (double.IsInfinity(size.Width))
-                    {
-                        isNeedLoadMore = (args.Index + 1) * size.Height <= viewportHeight;
-                    }
-                    else
-                    {
-                        var rowCount = args.Index / (viewportWidth / size.Width);
-                        isNeedLoadMore = rowCount * size.Height <= viewportHeight;
-                    }
-
-                    if (isNeedLoadMore)
-                    {
-                        RequestLoadMore?.Invoke(this, EventArgs.Empty);
-                    }
-                }
-            }
         }
     }
 }
