+++ src/App/Themes/Generic.xaml
@@ -11,5 +11,6 @@
         <ResourceDictionary Source="ms-appx:///Controls/App/CenterPopup/CenterPopup.xaml" />
         <ResourceDictionary Source="ms-appx:///Controls/App/IconTextBlock/IconTextBlock.xaml" />
         <ResourceDictionary Source="ms-appx:///Controls/Common/EmotiTextBlock/EmotiTextBlock.xaml" />
+        <ResourceDictionary Source="ms-appx:///Controls/Common/VerticalRepeaterView/VerticalRepeaterView.xaml" />
     </ResourceDictionary.MergedDictionaries>
 </ResourceDictionary>
